refactor: clean up login/join file structure into feature-based folders
